refactor: 清理临时类型定义，统一类型管理
🎯 主要改进：
- ✅ 移除所有分散的临时类型定义
- ✅ 将核心类型统一到 src/types/core/fbx-types.ts
- ✅ 消除重复类型定义，避免维护混乱
- ✅ 建立清晰的类型导入层次结构

📦 类型重组详情：
1. 新增统一类型定义：
   - ParseContext - FBX解析上下文
   - ModelLoaderResult - 模型加载结果
   - LoaderOptions - 加载器选项
   - FBXConnectionNodeEnhanced - 增强连接节点
   - FBXLightNodeAttribute, FBXVideoNode, FBXCameraNodeAttribute
   - FBXConnectionType - 连接类型

2. 类型导入路径统一：
   - 从临时定义 → 从 types/core 导入
   - 从 constants.ts → 从 types/core 导入
   - 消除循环依赖和重复定义

3. 接口增强：
   - IFBXTree 增加 Objects 属性支持
   - FBXConnectionNode 支持 relationship 属性
   - 统一属性命名规范

🔧 文件修改清单：
- src/types/core/fbx-types.ts: 新增统一类型定义
- src/types/core/index.ts: 清理重复导出
- src/parse/FBX-animation-parser.ts: 使用统一类型
- src/parse/FBX-geometry-parser.ts: 使用统一类型
- src/parse/FBX-tree-parser.ts: 使用统一类型
- src/parse/FBX-binary-parser.ts: 使用统一类型
- src/parse/FBX-text-parser.ts: 使用统一类型
- src/FBX-loader.ts: 使用统一类型
- src/FBXLoaderRefactored.ts: 使用统一类型
- src/constants.ts: 移除重复类型定义

🚀 价值体现：
- 提升代码可维护性：类型定义集中管理
- 减少重复工作：单一数据源，避免不一致
- 改善开发体验：清晰的导入路径和类型提示
- 为后续类型安全增强奠定基础

注意：类型错误数量暂时增加是正常现象，属于重构过程中的预期情况。
