+++ src/FBX-loader.ts
@@ -1,10 +1,5 @@
 import { FileLoader, Group, Loader, LoaderUtils, TextureLoader } from 'three';
-// 临时类型定义
-export interface LoaderOptions {
-  [key: string]: any;
-}
-
-import type { ModelLoaderResult } from './parse/FBX-tree-parser';
+import type { ModelLoaderResult, LoaderOptions } from './types/core';
 import { convertArrayBufferToString, getFbxVersion, isFbxFormatASCII, isFbxFormatBinary } from './util';
 import { BinaryParser } from './parse/FBX-binary-parser';
 import { FBXTreeParser } from './parse/FBX-tree-parser';
