+++ src/parse/FBX-animation-parser.ts
@@ -12,13 +12,7 @@ import {
   VectorKeyframeTrack,
 } from 'three';
 import { convertFBXTimeToSeconds, getEulerOrder } from './utils';
-import { FBXConnectionNode } from '../constants';
-
-// 临时定义 ParseContext 类型
-export interface ParseContext {
-  connections: Map<number, FBXConnectionNode>;
-  [key: string]: any;
-}
+import type { ParseContext } from '../types/core';
 
 interface AnimationCurve {
   id: number,
