+++ src/parse/FBX-tree-parser.ts
@@ -50,37 +50,15 @@ import type {
   FBXTextureNode,
   RawBone,
 } from '../constants';
-import { FBXConnectionNode } from '../constants';
-
-// 临时类型定义
-export interface ParseContext {
-  connections: Map<number, FBXConnectionNode>;
-  [key: string]: any;
-}
-
-export interface ModelLoaderResult {
-  scene: any;
-  animations: any[];
-  modelInfo: any;
-}
-
-export interface BaseInfo {
-  id: number;
-  name: string;
-  type?: string;
-}
-
-export interface FBXLightNodeAttribute {
-  [key: string]: any;
-}
-
-export interface FBXVideoNode {
-  [key: string]: any;
-}
-
-export interface FBXCameraNodeAttribute {
-  [key: string]: any;
-}
+import type { 
+  ParseContext, 
+  ModelLoaderResult, 
+  BaseInfo, 
+  FBXLightNodeAttribute, 
+  FBXVideoNode, 
+  FBXCameraNodeAttribute,
+  FBXConnectionNode 
+} from '../types/core';
 import { generateTransform, getEulerOrder } from './utils';
 import { AnimationParser } from './FBX-animation-parser';
 import { GeometryParser } from './FBX-geometry-parser';
